backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
psycopg2-binary
sqlalchemy
openpyxl
ta 